zstd: address review — dedup shims, -arch generation, asm-vs-nonasm fuzz
Responds to maintainer review on the arm64 decoder PR:

- Deduplicate the asm Go wrappers: shared seqdec_asm.go holds the context
  structs, error codes and the decode/decodeSync/executeSimple bodies; each
  arch keeps only its //go:noescape decls and a small dispatch helper (amd64
  selects the BMI2 / 56-bit / safe variant, arm64 the 56-bit / safe variant).
  The fse Go wrapper is identical across arches and collapses into one
  fse_decoder_asm.go (per-arch fse_decoder_{amd64,arm64}.go removed).

- Generate both architectures via avo's new -arch flag: one
  `go run gen.go -out ../seqdec.s -arch amd64,arm64` writes seqdec_amd64.s and
  seqdec_arm64.s from a single compile, so neither file's header references the
  other architecture.

- FuzzDecodeAll now cross-checks the assembly decoder against a non-assembly
  path (BMI2 disabled on amd64) for byte-identical output; run with -tags=noasm
  over the shared corpus to cover assembly-vs-generic on arm64.
